Add new update script
